generalize: manual talairach file is now called talairach_man.xfm instead of RB_AVERAGE3new.xfm; if manual talairach not found, then try to use the auto-generated one; allow specifying a name prefix for the atlas to use instead of RB

+++ scripts/rebuild_gca_atlas.csh
@@ -18,16 +18,18 @@
 # each subject's mri directory: nu.mgz, brain.mgz, nu_noneck.mgz and
 # seg_edited.mgz, where seg_edited.mgz is the manually labelled volume
 # from which the training data is derived.
-# The final atlas is stored as $SUBJECTS_DIR/average/RB_all_`date +%F`.gca
+# The final atlas is stored as:
+#  $SUBJECTS_DIR/average/${GCA_PRE}_all_`date +%F`.gca
+# where ${GCA_PRE} specifies the prefix (default is 'RB').
 # "nu_noneck.mgz" is needed for every subject to build the gca with skull.
 #
 # Original author: Xiao Han
 # CVS Revision Info:
 #    $Author: nicks $
-#    $Date: 2008/03/13 15:54:43 $
-#    $Revision: 1.14.2.1 $
+#    $Date: 2008/10/08 19:21:33 $
+#    $Revision: 1.14.2.2 $
 #
-# Copyright (C) 2002-2007,
+# Copyright (C) 2002-2008,
 # The General Hospital Corporation (Boston, MA).
 # All rights reserved.
 #
@@ -41,13 +43,18 @@
 #
 
 
-set VERSION='$Id: rebuild_gca_atlas.csh,v 1.14.2.1 2008/03/13 15:54:43 nicks Exp $';
+set VERSION='$Id: rebuild_gca_atlas.csh,v 1.14.2.2 2008/10/08 19:21:33 nicks Exp $';
 
 #set echo=1
 
 # these are the subjects to use in training.
 # one of them, the target, set to ONE_SUBJECT, should be declared is this file:
 source ${SUBJECTS_DIR}/scripts/subjects.csh
+# also, GCA_PRE can optionally be defined in subjects.csh, to override the
+# default name prefix of 'RB'.
+if ( ! $?GCA_PRE) then
+  set GCA_PRE=(RB)
+endif
 
 # created GCA files, and log file, get this date appended
 set DATE=(`date +%F`)
@@ -103,7 +110,7 @@ set SEG_VOL=(seg_edited.mgz) # filename (symlink) for manual segmentation
 set ORIG_VOL=(nu.mgz)
 set MASK_VOL=(brain.mgz) # filename for brain mask
 set T1_NONECK=(nu_noneck.mgz) # file to build the atlas gca_with_skull
-set TAL_MAN=(RB_AVERAGE3new.xfm) # optional manual tal registration file
+set TAL_MAN=(talairach_man.xfm) # optional manual tal registration file
 set INPUTS=(${SEG_VOL} ${ORIG_VOL} ${MASK_VOL} ${T1_NONECK})
 set ALL_SUBJS=(${SUBJECTS} ${ONE_SUBJECT})
 foreach subject (${ALL_SUBJS}) # check for existence of required inputs
@@ -121,10 +128,7 @@ foreach subject (${ALL_SUBJS}) # check for existence of required inputs
         endif
     end
 end
-if ( ! -e ${SUBJECTS_DIR}/${ONE_SUBJECT}/mri/transforms/${TAL_MAN} ) then
-    echo "Missing ${SUBJECTS_DIR}/${ONE_SUBJECT}/mri/transforms/${TAL_MAN}!"
-    exit 1
-endif
+
 
 
 ##########################################################################
@@ -136,9 +140,9 @@ endif
 # Outputs:
 #
 mkdir -p ${SUBJECTS_DIR}/average
-set GCA=(${SUBJECTS_DIR}/average/RB_all_${DATE}.gca)
-set GCA_ONE=(${SUBJECTS_DIR}/average/RB_one_${DATE}.gca)
-set GCA_SKULL=(${SUBJECTS_DIR}/average/RB_all_withskull_${DATE}.gca)
+set GCA=(${SUBJECTS_DIR}/average/${GCA_PRE}_all_${DATE}.gca)
+set GCA_ONE=(${SUBJECTS_DIR}/average/${GCA_PRE}_one_${DATE}.gca)
+set GCA_SKULL=(${SUBJECTS_DIR}/average/${GCA_PRE}_all_withskull_${DATE}.gca)
 
 set LTA_ONE=(talairach_one.lta)
 set M3D_ONE=(talairach_one.m3z)
@@ -164,7 +168,7 @@ if ("$1" == "2") goto stage2
 echo "$canorm each subject using its ${SEG_VOL}, producing ${T1_VOL}..."
 foreach subject (${SUBJECTS})
     set mridir=(${SUBJECTS_DIR}/$subject/mri)
-    if(-e $mridir/${T1_VOL}) rm -f $mridir/${T1_VOL}
+    if($RunIt & -e $mridir/${T1_VOL}) rm -f $mridir/${T1_VOL}
 
     set cmd=($canorm -mask $mridir/${MASK_VOL})
     set cmd=($cmd -seg $mridir/${SEG_VOL})
@@ -199,8 +203,13 @@ if(-e ${SUBJECTS_DIR}/$ONE_SUBJECT/mri/transforms/${TAL_MAN}) then
     echo "Using manual talairach registration ${TAL_MAN}..." >>& $LF
     set MAN_TAL=(-xform ${TAL_MAN})
 else
+  if(-e ${SUBJECTS_DIR}/$ONE_SUBJECT/mri/transforms/talairach.xfm) then
+    echo "Using talairach registration talairach.xfm..." >>& $LF
+    set MAN_TAL=(-xform talairach.xfm)
+  else
     echo "Initial talairach registration unavailable (no xform)..." >>& $LF
     set MAN_TAL=
+  endif
 endif
 echo "$train using one subject: $ONE_SUBJECT, producing ${GCA_ONE}..."
 set cmd=($train -prior_spacing 2 -node_spacing 8 -mask ${MASK_VOL})
@@ -230,27 +239,11 @@ foreach subject (${SUBJECTS})
     set mridir=(${SUBJECTS_DIR}/$subject/mri)
     mkdir -p $mridir/transforms
     set tdir=($mridir/transforms)
-    if ( -e $tdir/${LTA_ONE} ) rm -f $tdir/${LTA_ONE}
+    if ( $RunIt & -e $tdir/${LTA_ONE} ) rm -f $tdir/${LTA_ONE}
 
     set cmd=($emreg -mask $mridir/${MASK_VOL} $mridir/${ORIG_VOL})
     set cmd=($cmd ${GCA_ONE} $mridir/transforms/${LTA_ONE})
-    echo $cmd >>& $LF
-    if ($RunIt) pbsubmit ${PBCONF} -c "$cmd"
-end
-echo "\n\n" >>& $LF
-# wait until the registration is done
-foreach subject (${SUBJECTS})
-    set TEST = 0
-    if ($RunIt) set TEST = 1
-    while($TEST)
-        if(-e ${SUBJECTS_DIR}/$subject/mri/transforms/${LTA_ONE}) then
-            set TEST = 0
-        else
-            sleep 30
-        endif
-    end
-    echo "\t...finished $emreg on subject $subject" >>& $LF
-end
+
 ##########################################################################
 
 
@@ -259,31 +252,12 @@ end
 #
 # normalization by GCA_ONE
 #
-echo "$canorm each subject using one-subject GCA, producing ${T1_VOL}..."
-foreach subject (${SUBJECTS})
-    set mridir=(${SUBJECTS_DIR}/$subject/mri)
-    if ( -e $mridir/${T1_VOL} ) rm -f $mridir/${T1_VOL}
+    if ($RunIt &  -e $mridir/${T1_VOL} ) rm -f $mridir/${T1_VOL}
 
-    set cmd=($canorm -mask $mridir/${MASK_VOL} $mridir/${ORIG_VOL})
+    set cmd=($cmd; $canorm -mask $mridir/${MASK_VOL} $mridir/${ORIG_VOL})
     set cmd=($cmd ${GCA_ONE} $mridir/transforms/${LTA_ONE})
     set cmd=($cmd $mridir/${T1_VOL})
-    echo $cmd >>& $LF
-    if ($RunIt) pbsubmit ${PBCONF} -c "$cmd"
-end
-echo "\n\n" >>& $LF
-# wait until the normalization is done
-foreach subject (${SUBJECTS})
-    set TEST = 0
-    if ($RunIt) set TEST = 1
-    while($TEST)
-        if(-e ${SUBJECTS_DIR}/$subject/mri/${T1_VOL}) then
-            set TEST = 0
-        else
-            sleep 30
-        endif
-    end
-    echo "\t...finished $canorm on subject $subject" >>& $LF
-end
+
 ##########################################################################
 
 
@@ -292,12 +266,9 @@ end
 #
 # CA_registration by GCA_ONE
 #
-echo "$careg each subject to one-subj GCA, producing transforms/${M3D_ONE}..."
-foreach subject (${SUBJECTS})
-    set mridir=(${SUBJECTS_DIR}/$subject/mri)
-    if ( -e  $mridir/transforms/${M3D_ONE}) rm -f $mridir/transforms/${M3D_ONE}
+    if ($RunIt &  -e  $mridir/transforms/${M3D_ONE}) rm -f $mridir/transforms/${M3D_ONE}
 
-    set cmd=($careg -align-after -smooth 1.0 -levels 2 -mask $mridir/${MASK_VOL})
+    set cmd=($cmd; $careg -align-after -smooth 1.0 -levels 2 -mask $mridir/${MASK_VOL})
     set cmd=($cmd -T $mridir/transforms/${LTA_ONE} $mridir/${T1_VOL})
     set cmd=($cmd ${GCA_ONE} $mridir/transforms/${M3D_ONE})
     echo $cmd >>& $LF
@@ -334,7 +305,7 @@ stage2:
 #
 # TRAIN FROM SEGMENTED_SUBJECTS USING M3D_ONE
 #
-if (-e ${GCA} ) rm -f ${GCA}
+if ($RunIt & -e ${GCA} ) rm -f ${GCA}
 
 echo "$train using all subjects, using ${M3D_ONE}, producing ${GCA}..."
 set cmd=($train -prior_spacing 2 -node_spacing 4 -mask ${MASK_VOL})
@@ -354,6 +325,14 @@ echo "\t...finished $train, produced ${GCA}" >>& $LF
 ##########################################################################
 
 
+##########################################################################
+
+
+#
+# may want to restart here
+# to start-up here, use '3' as the input arg. default is run both stages.
+#
+stage3:
 
 ##########################################################################
 #
@@ -364,28 +343,11 @@ echo "$emreg each subject to GCA, producing transforms/${LTA}..."
 foreach subject (${SUBJECTS})
     set mridir=(${SUBJECTS_DIR}/$subject/mri)
     set tdir=($mridir/transforms)
-    if ( -e $tdir/${LTA} ) rm -f $tdir/${LTA}
+    if ($RunIt & -e $tdir/${LTA} ) rm -f $tdir/${LTA}
 
     set cmd=($emreg -mask $mridir/${MASK_VOL} $mridir/${ORIG_VOL})
     set cmd=($cmd ${GCA})
     set cmd=($cmd $mridir/transforms/${LTA})
-    echo $cmd >>& $LF
-    if ($RunIt) pbsubmit ${PBCONF} -c "$cmd"
-end
-echo "\n\n" >>& $LF
-# wait until the registration is done
-foreach subject (${SUBJECTS})
-    set TEST = 0
-    if ($RunIt) set TEST = 1
-    while($TEST)
-        if(-e ${SUBJECTS_DIR}/$subject/mri/transforms/${LTA}) then
-            set TEST = 0
-        else
-            sleep 30
-        endif
-    end
-    echo "\t...finished $emreg on subject $subject" >>& $LF
-end
 ##########################################################################
 
 
@@ -394,31 +356,11 @@ end
 #
 # normalization by GCA
 #
-echo "$canorm each subject using GCA, producing ${T1_VOL}..."
-foreach subject (${SUBJECTS})
-    set mridir=(${SUBJECTS_DIR}/$subject/mri)
-    if ( -e $mridir/${T1_VOL} ) rm -f $mridir/${T1_VOL}
+    if ($RunIt &  -e $mridir/${T1_VOL} ) rm -f $mridir/${T1_VOL}
 
-    set cmd=($canorm -mask $mridir/${MASK_VOL} $mridir/${ORIG_VOL})
+    set cmd=($cmd ; $canorm -mask $mridir/${MASK_VOL} $mridir/${ORIG_VOL})
     set cmd=($cmd ${GCA} $mridir/transforms/${LTA})
     set cmd=($cmd $mridir/${T1_VOL})
-    echo $cmd >>& $LF
-    if ($RunIt) pbsubmit ${PBCONF} -c "$cmd"
-end
-echo "\n\n" >>& $LF
-# wait until the normalization is done
-foreach subject (${SUBJECTS})
-    set TEST = 0
-    if ($RunIt) set TEST = 1
-    while($TEST)
-        if(-e ${SUBJECTS_DIR}/$subject/mri/${T1_VOL}) then
-            set TEST = 0
-        else
-            sleep 30
-        endif
-    end
-    echo "\t...finished $canorm on subject $subject" >>& $LF
-end
 ##########################################################################
 
 
@@ -427,12 +369,9 @@ end
 #
 # CA_registration by GCA
 #
-echo "$careg each subject to GCA, producing transforms/${M3D}..."
-foreach subject (${SUBJECTS})
-    set mridir=(${SUBJECTS_DIR}/$subject/mri)
-    if ( -e $mridir/transforms/${M3D} ) rm -f $mridir/transforms/${M3D}
+    if ($RunIt & -e $mridir/transforms/${M3D} ) rm -f $mridir/transforms/${M3D}
 
-    set cmd=($careg -align-after -smooth 1.0 -mask $mridir/${MASK_VOL})
+    set cmd=($cmd ; $careg -align-after -smooth 1.0 -mask $mridir/${MASK_VOL})
     set cmd=($cmd -T $mridir/transforms/${LTA} $mridir/${T1_VOL})
     set cmd=($cmd ${GCA} $mridir/transforms/${M3D})
     echo $cmd >>& $LF
@@ -460,7 +399,7 @@ end
 #
 # RETRAIN GCA USING M3D
 #
-rm -f ${GCA}
+if ($RunIt) rm -f ${GCA}
 
 echo "$train, using ${M3D}, producing ${GCA}..."
 set cmd=($train -prior_spacing 2 -node_spacing 4 -mask ${MASK_VOL})
@@ -478,7 +417,7 @@ while($TEST)
 end
 echo "\t...finished $train, produced final $GCA" >>& $LF
 
-rm -f ${GCA_SKULL}
+if ($RunIt) rm -f ${GCA_SKULL}
 
 echo "$train, using ${LTA}, producing ${GCA_SKULL}..."
 set cmd=($train -prior_spacing 2 -node_spacing 4)
@@ -540,7 +479,7 @@ foreach subject (${SUBJECTS})
     echo "\t...subject $subject has orig.mgz"
 end
 foreach subject (${SUBJECTS})
-    rm -f ${SUBJECTS_DIR}/$subject/mri/brain.mgz
+    if ($RunIt) rm -f ${SUBJECTS_DIR}/$subject/mri/brain.mgz
     set cmd=(recon-all -s $subject)
     set cmd=($cmd -nuintensitycor)
     set cmd=($cmd -talairach)
